Write more advanced testbed
Check for equivalence of css, not merely that the input and output files
are identical. This means that tests are not sensitive to formatting,
order, or splitting of css between rules.

One case that could be improved is:
    .foo, .bar { display: none }
    vs
    .foo { display: none }
    .bar { display: none }

Also, failure to compile is taken as a failed unit test. I should
include a mechanism to test for desired failure cases.