Improve mixins
Allow any selector to be used to specify an included mixin, not just
a class selector.
Allow multiple mixins to be matched. An error is raised only if no
mixins were matched by name. (No error is raised if there is no match
because of guard, arity, or pattern mismatch.)
Fix bug which allowed zero-length string literals to be parsed.
(Which causes problems with statements like `.mixin();`, because we
parse one argument (a string of length 0), not 0 arguments.